feat(mcp-proxy): binary topology — obsigna-mcp + obsigna mcp run (ADR-0033)
