+++ mcp-proxy/cmd/obsigna-mcp/cli_test.go
