+++ mcp-proxy/cmd/obsigna-mcp/dirperms.go
