+++ mcp-proxy/cmd/obsigna-mcp/dirperms_other.go
