+++ mcp-proxy/cmd/obsigna-mcp/dirperms_test.go
