+++ mcp-proxy/cmd/obsigna-mcp/dirperms_unix.go
