+++ mcp-proxy/cmd/obsigna-mcp/dirperms_unix_test.go
