+++ mcp-proxy/cmd/obsigna-mcp/emitter_integration_test.go
