+++ mcp-proxy/cmd/obsigna-mcp/shutdown_test.go
